Move utilities to `bld_collection_utils.py`
