+++ intake_esm/cmip.py
@@ -9,6 +9,8 @@ from intake_esm import aggregate, config
 from intake_esm.collection import Collection, docstrings
 from intake_esm.source import BaseSource
 
+from .bld_collection_utils import _extract_attr_with_regex, _reverse_filename_format
+
 
 class CMIP5Collection(Collection):
 
@@ -43,16 +45,14 @@ class CMIP5Collection(Collection):
             '{variable}_{mip_table}_{model}_{experiment}_{ensemble_member}_{temporal_subset}.nc'
         )
         gridspec_template = '{variable}_{mip_table}_{model}_{experiment}_{ensemble_member}.nc'
-        f = CMIP5Collection._reverse_filename_format(
+        f = _reverse_filename_format(
             file_basename, filename_template=filename_template, gridspec_template=gridspec_template
         )
         fileparts.update(f)
 
-        frequency = CMIP5Collection._extract_attr_with_regex(
-            filepath, regex=freq_regex, strip_chars='/'
-        )
-        realm = CMIP5Collection._extract_attr_with_regex(filepath, regex=realm_regex)
-        version = CMIP5Collection._extract_attr_with_regex(filepath, regex=version_regex) or 'v0'
+        frequency = _extract_attr_with_regex(filepath, regex=freq_regex, strip_chars='/')
+        realm = _extract_attr_with_regex(filepath, regex=realm_regex)
+        version = _extract_attr_with_regex(filepath, regex=version_regex) or 'v0'
         fileparts['frequency'] = frequency
         fileparts['modeling_realm'] = realm
         fileparts['version'] = version
@@ -104,12 +104,12 @@ class CMIP6Collection(Collection):
             '{variable_id}_{table_id}_{source_id}_{experiment_id}_{member_id}_{grid_label}.nc'
         )
 
-        f = CMIP6Collection._reverse_filename_format(
+        f = _reverse_filename_format(
             file_basename, filename_template=filename_template, gridspec_template=gridspec_template
         )
         fileparts.update(f)
         version_regex = r'v\d{4}\d{2}\d{2}|v\d{1}'
-        version = CMIP6Collection._extract_attr_with_regex(filepath, regex=version_regex) or 'v0'
+        version = _extract_attr_with_regex(filepath, regex=version_regex) or 'v0'
         fileparts['version'] = version
 
         return fileparts
