+++ intake_esm/collection.py
@@ -1,18 +1,13 @@
-import datetime
 import os
-import re
 from abc import ABC, abstractclassmethod
-from glob import glob
-from pathlib import Path
 
 import docrep
 import pandas as pd
-import pkg_resources
 import xarray as xr
-from intake.source.utils import reverse_format
 from tqdm.autonotebook import tqdm
 
 from . import config
+from .bld_collection_utils import make_attrs
 from .storage import StorageResource
 
 docstrings = docrep.DocstringProcessor()
@@ -77,8 +72,8 @@ class Collection(ABC):
             df_i = self.assemble_file_list(data_source, data_source_attrs, self.exclude_patterns)
             dfs.update(df_i)
 
-        self.df = self._finalize_build(dfs)
-        self._ds = self.df.reset_index(drop=True).to_xarray()
+        self._ds = self._finalize_build(dfs).reset_index(drop=True).to_xarray()
+
         attrs = make_attrs(
             attrs={
                 'collection_spec': self.collection_spec,
@@ -87,8 +82,8 @@ class Collection(ABC):
             }
         )
         self._ds.attrs = attrs
+        self._persist_db_file()
         print(self._ds)
-        self.persist_db_file()
 
     def assemble_file_list(self, data_source, data_source_attrs, exclude_patterns=[]):
         """ Assemble file listing for data sources into Pandas dataframes.
@@ -135,42 +130,6 @@ class Collection(ABC):
             for key, value in extra_attrs.items():
                 df[key] = value
         return df
-
-    @staticmethod
-    def _extract_attr_with_regex(input_str, regex, strip_chars=None):
-        pattern = re.compile(regex)
-        match = re.findall(pattern, input_str)
-        if match:
-            match = max(match, key=len)
-            if strip_chars:
-                match = match.strip(strip_chars)
-
-            else:
-                match = match.strip()
-
-            return match
-
-        else:
-            return None
-
-    @staticmethod
-    def _reverse_filename_format(file_basename, filename_template=None, gridspec_template=None):
-        """
-        Uses intake's ``reverse_format`` utility to reverse the string method format.
-
-        Given format_string and resolved_string, find arguments
-        that would give format_string.format(arguments) == resolved_string
-        """
-        try:
-            return reverse_format(filename_template, file_basename)
-        except ValueError:
-            try:
-                return reverse_format(gridspec_template, file_basename)
-            except:
-                print(
-                    f'Failed to parse file: {file_basename} using patterns: {filename_template} and {gridspec_template}'
-                )
-                return {}
 
     def _assemble_collection_df_files(
         self, resource_key, resource_type, direct_access, filelist, urlpath=None
@@ -227,8 +186,8 @@ class Collection(ABC):
 
         Returns
         --------
-        df : pandas.DataFrame
-            Cleaned pandas dataframe containing all entries
+        ds : pandas.DataFrame
+            Cleaned dataframe containing all entries
 
         Notes
         -----
@@ -265,7 +224,7 @@ class Collection(ABC):
         return exclude_patterns
 
     def _validate(self):
-        """Checks that collection columns are properly defined in `confi.yaml` file.
+        """Checks that collection columns are properly defined in `config.yaml` file.
         """
         for req_col in config.get('collections')[self.collection_spec['collection_type']][
             'required-columns'
@@ -275,10 +234,10 @@ class Collection(ABC):
                     f"Missing required column: {req_col} for {self.collection_spec['collection_type']} in {config.PATH}"
                 )
 
-    def persist_db_file(self):
+    def _persist_db_file(self):
         """ Persist built collection database to disk.
         """
-        if not self.df.empty:
+        if len(self._ds) > 0:
             print(
                 f"Persisting {self.collection_spec['name']} at : {os.path.abspath(self.collection_db_file)}"
             )
@@ -294,95 +253,4 @@ class Collection(ABC):
             )
 
         else:
-            print(f"{self.df} is an empty dataframe. It won't be persisted to disk.")
-
-
-def _get_built_collections():
-    """Loads built collections in a dictionary with key=collection_name, value=collection_db_file_path"""
-    try:
-        db_dir = Path(config.get('database-directory'))
-        cc = db_dir.glob('*.nc')
-        collections = {}
-        for f in cc:
-            name = f.stem
-            fullpath = f.absolute()
-            collections[name] = fullpath
-        return collections
-    except Exception as e:
-        raise e
-
-
-def _open_collection(collection_name):
-    """ Open an ESM collection"""
-    collections = _get_built_collections()
-
-    ds = xr.open_dataset(collections[collection_name], engine='netcdf4')
-    # ds['direct_access'] = ds['direct_access'].astype(bool)
-    collection_type = ds.attrs['collection_type']
-    collection_name = ds.attrs['name']
-    return ds.to_dataframe(), collection_name, collection_type
-
-
-def get_subset(collection_name, query, order_by=None):
-    """ Get a subset of collection entries that match a query """
-    import numpy as np
-
-    df, _, collection_type = _open_collection(collection_name)
-
-    condition = np.ones(len(df), dtype=bool)
-
-    for key, val in query.items():
-        if isinstance(val, list):
-            condition_i = np.zeros(len(df), dtype=bool)
-            for val_i in val:
-                condition_i = condition_i | (df[key] == val_i)
-            condition = condition & condition_i
-
-        elif val is not None:
-            condition = condition & (df[key] == val)
-    query_results = df.loc[condition]
-
-    if order_by is None:
-        order_by = config.get('collections')[collection_type]['order-by-columns']
-
-    query_results = query_results.sort_values(by=order_by, ascending=True)
-
-    return query_results
-
-
-def make_attrs(attrs=None):
-    """Make standard attributes to attach to xarray datasets (collections).
-    Parameters
-    ----------
-    attrs : dict (optional)
-        Additional attributes to add or overwrite
-    Returns
-    -------
-    dict
-        attrs
-    """
-
-    import intake_xarray
-    import intake
-    import json
-
-    default_attrs = {
-        'created_at': datetime.datetime.utcnow().isoformat(),
-        'intake_esm_version': pkg_resources.get_distribution('intake_esm').version,
-    }
-
-    upstream_deps = [intake, intake_xarray]
-    for dep in upstream_deps:
-        dep_name = dep.__name__
-        try:
-            version = pkg_resources.get_distribution(dep_name).version
-            default_attrs[f'{dep_name}_version'] = version
-        except pkg_resources.DistributionNotFound:
-            if hasattr(dep, '__version__'):
-                version = dep.__version__
-                default_attrs[f'{dep_name}_version'] = version
-    if attrs is not None:
-        if 'collection_spec' in attrs:
-            attrs['collection_spec'] = json.dumps(attrs['collection_spec'])
-        default_attrs.update(attrs)
-    return default_attrs
+            print(f"{self._ds} is an empty dataset. It won't be persisted to disk.")
