+++ intake_esm/source.py
@@ -3,7 +3,7 @@ import xarray as xr
 from tqdm.autonotebook import tqdm
 
 from . import aggregate
-from .collection import get_subset
+from .bld_collection_utils import get_subset
 from .storage import _ensure_file_access
